refactor(services): 견적서 리스트/상세 조회 컴포넌트화 및 디자인시스템 적용
