+++ packages/services/src/estimate/components/estimate/add-input/index.styles.ts
@@ -1,8 +1,7 @@
+import { theme } from '@daengle/design-system';
 import { css } from '@emotion/react';
 
 export const addTitle = css`
-  font-size: 14px;
-  color: #979797;
   padding: 24px 18px 14px;
 `;
 
@@ -12,15 +11,14 @@ export const inputSection = css`
   justify-content: center;
   align-items: center;
   padding: 0px 18px;
-  margin-bottom: 16px;
+  margin-bottom: 14px;
 `;
 
 export const textarea = css`
   width: 100%;
   min-height: 80px;
   padding: 8px;
-  font-size: 14px;
-  color: #000000;
+  ${theme.typo.body4}
   background-color: #f0f0f0;
   border-radius: 10px;
   resize: none;
