+++ packages/services/src/estimate/components/estimate/add-input/index.tsx
@@ -1,3 +1,4 @@
+import { Text } from '@daengle/design-system';
 import { addTitle, inputSection, textarea } from './index.styles';
 
 interface Props {
@@ -9,7 +10,11 @@ interface Props {
 const AddInput = ({ title, placeholder, height }: Props) => {
   return (
     <div>
-      <div css={addTitle}>{title}</div>
+      <div css={addTitle}>
+        <Text typo="body4" color="gray400">
+          {title}
+        </Text>
+      </div>
       <section css={inputSection}>
         <textarea
           css={[textarea, height && { minHeight: `${height}px` }]}
