+++ packages/services/src/estimate/components/estimate/pet-profile/index.tsx
@@ -1,3 +1,5 @@
+import { TextButton, Text } from '@daengle/design-system';
+import { ButtonTextButtonArrow } from '@daengle/design-system/icons';
 import {
   wrapper,
   profile,
@@ -6,40 +8,42 @@ import {
   detail,
   labelWrapper,
   valueWrapper,
-  label,
-  value,
-  detailButton,
 } from './index.styles';
 
 interface Props {
   image: string;
   name: string;
-  attributes: { label: string; value: string }[];
+  attributes: (string | number)[];
 }
 
 const PetDetails = ({ image, name, attributes }: Props) => {
+  const labels = ['나이', '몸무게', '특이사항'];
   return (
     <div css={wrapper}>
       <div css={profile}>
         <img src={image} alt={`${name} 프로필`} css={imageUrl} />
-        <div css={petName}>{name}</div>
+        <Text typo="subtitle3" css={petName}>
+          {name}
+        </Text>
       </div>
       <div css={detail}>
         <div css={labelWrapper}>
-          {/* api 연동 시 key 값 변경 */}
-          {attributes.map((attribute, index) => (
-            <div css={label} key={`label-${index}`}>
-              {attribute.label}
-            </div>
+          {labels.map((label, index) => (
+            <Text typo="body9" color="gray500" key={`label-${index}`}>
+              {label}
+            </Text>
           ))}
-          <div css={detailButton}>자세히보기 {'>'}</div>
+          <TextButton icons={{ suffix: <ButtonTextButtonArrow width={'6px'} /> }}>
+            <Text color="gray200" typo="body9">
+              자세히보기
+            </Text>
+          </TextButton>
         </div>
         <div css={valueWrapper}>
-          {/* api 연동 시 key 값 변경 */}
           {attributes.map((attribute, index) => (
-            <div css={value} key={`value-${index}`}>
-              {attribute.value}
-            </div>
+            <Text typo="body9" key={`value-${index}`}>
+              {attribute}
+            </Text>
           ))}
         </div>
       </div>
